+++ src/App.js
@@ -1,6 +1,7 @@
 import './App.css';
 import Header from './components/header/header';
 import HomePage from './components/body/homepage';
+import SignupForm from './components/signUp/SignupForm';
 import React, {Fragment} from 'react';
 function App() {
   return (
@@ -8,8 +9,8 @@ function App() {
       <Header className = "header"/>
       <body className='mainbody'>
         <HomePage/>
-        {/* <SignupForm className="login"/>
-        <LoginForm className="login"/> */}
+        {/* <SignupForm className="login"/> */}
+        {/* <LoginForm className="login"/> */}
       </body>
     </Fragment>
   );
