+++ src/components/signUp/signup.css
@@ -1,0 +1,175 @@
+* {
+    /* margin: 0; */
+    box-sizing: border-box;
+    font-family: "CustomFont";
+}
+
+/* @font-face {
+    font-family: "CustomFont";
+    src: url("../fonts/RolexLight.woff2");
+} */
+
+/* --------------------------- MOBILE VERSION ------------------------------ */
+/* .page {
+    height: 100vh;
+    width: 100%;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    font-size: 0.75em;
+} */
+
+.lgimg{
+    position: absolute;
+    top:0;
+    left:0;
+    margin-top: 2px;
+    margin-left: 10px;
+    /* margin-left: 20px; */
+}
+.signupcontainer{
+    width: 100%;
+    height: 100%;
+    /* background-color: aqua; */
+    /* height: 32rem; */
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    position: fixed;
+    top: 0;
+    bottom: 0;
+    left: 0;
+    right: 0;
+    backdrop-filter: blur(3px);
+    /* border:2px solid red; */
+   
+}
+.signupcover {
+    width: 400px;
+    height: 550px;
+    margin-bottom: 20px;
+    display: flex;
+    flex-direction: column;
+    background-color: white;
+    border-radius: 20px;
+}
+.signupheading{
+    text-align: center;
+    /* padding-right: 2em; */
+    /* width: 600px; */
+    font-family: sans-serif;
+
+    color: #003049;
+    /* font-weight: 600; */
+    font-size: 25px;
+    
+   margin-top:10px;
+   
+}
+.subcon{
+    
+   margin: auto;
+   margin-top: -3px;
+   width: 65%;
+   font-size: 12px;
+}
+.fullname{
+   padding-top: 2px;
+    
+    /* background-color: #2C2929; */
+}
+
+input {
+    border: 1px solid rgb(194, 192, 192);
+    /* background-color: rgb(229, 226, 226); */
+    width: 100%;
+    height: 2.2em;
+    border-radius: 0.25em;
+    /* text-align: center; */
+    padding-left: 1.5em;
+    font-family: sans-serif;
+    /* font-size: 0.75em; */
+    /* margin-top: 2px; */
+}
+
+input:focus {
+    outline-color:rgb(80, 74, 74);
+}
+
+.signup-btn {
+    width: 100%;
+    height: 3em;
+    background-color: #003049;;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    transition: 0.5s;
+    border-radius: 0.25em;
+    cursor: pointer;
+    margin-top: 2.5em;
+    color: white;
+    font-weight: 600;
+    line-height: 1.5em;
+    font-family: sans-serif;    
+}
+
+.login-btn:hover { transform: scale(1.05);}
+
+.text {
+    position: relative;
+    bottom: 0.5%;
+}
+
+.hide {
+    position: absolute;
+    transform: scale(0);
+}
+
+.login-popup {
+    position: absolute;
+    height: 30em;
+    width: 30em;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    border-radius: 0.25em;
+    background-color: black;
+    color: white;
+    bottom: 0;
+    transform: scale(1);
+    transition: 1.5s;
+}
+
+.subcon p{
+    font-family: sans-serif;
+    /* height: max-content; */
+    /* line-height: 27px;  */
+    color: #2C2929;
+    font-weight: 550;
+    /* border: 2px solid brown; */
+    /* margin-bottom: 0.4em; */
+}
+.forgot{
+    text-align: right;
+    font-weight: 600;
+    color: rgba(20, 102, 144, 1);
+    font-family: sans-serif;
+    font-size: small;
+    margin-top: 0.2em;
+}
+.signUp{
+    /* background-color: blue; */
+    margin-top: 10px;
+    width: 100%;
+    font-size: 15px;
+    /* text-align: center; */
+    font-family: sans-serif;
+    /* font-weight:bold; */
+}
+.signUp a{
+    text-decoration: none;
+    color: rgba(0, 48, 73, 1);
+    font-weight: bold;
+    font-family: sans-serif;
+}
